Add quirk to allow downloads on apple.com subdomains https://bugs.webkit.org/show_bug.cgi?id=281237 rdar://136805443
Reviewed by Chris Dumez.

As lamentable as it is, there is an internal apple.com subdomain that provides downloads
and has CSP that prevents downloads.  This worked until 283667@main made our behavior more
correct.  The downloads do not work in Chrome or Firefox.  The website owner is working
to resolve the issue, but until it is resolved I have been tasked with making the downloads
continue to work in Safari even though the HTTP header clearly and explicitly prevents downloads.

* Source/WebCore/loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::continueAfterContentPolicy):

Canonical link: https://commits.webkit.org/284999@main