Prevent use of HashMap::get() / HashMap::take() when the value type is an ObjectIdentifier https://bugs.webkit.org/show_bug.cgi?id=281306
Reviewed by NOBODY (OOPS!).

An ObjectIdentifier cannot be invalid/empty, we normally use `std::optional<ObjectIdentifier>`
or `Markable<ObjectIdentifier>` to represent an invalid/empty identifier. The only way to
construct an "empty" ObjectIdentifier is via `HashTraits<ObjectIdentifier>::emptyValue()`,
which HashMap relies on internally. As long as the HashMap is only using this special value
internally, this is fine. However, HashMap::get() and HashMap::take() will return this
invalid "empty" value if the key is not present. This was very error-prone for
ObjectIdentifier since the client would receive an invalid ObjectIdentifier and have no way
to check if the ObjectIdentifier is valid or not (since we assume an ObjectIdentifier is
always valid).

To address the issue, we now forbid the use of HashMap::get() / HashMap::take() when the
value type is an ObjectIdentifier and force the client to use `HashMap::getOptional()` /
`HashMap::takeOptional()` instead, which returns a std::optional.

* Source/WTF/wtf/HashMap.h:
(WTF::Y>::get const):
(WTF::Y>::inlineGet const):
(WTF::Y>::take):
(WTF::Y>::takeOptional):
* Source/WTF/wtf/HashSet.h:
(WTF::W>::take):
* Source/WTF/wtf/HashTraits.h:
* Source/WTF/wtf/ObjectIdentifier.h:
* Source/WebCore/workers/service/server/SWServer.cpp:
(WebCore::SWServer::matchAll):
* Source/WebKit/NetworkProcess/storage/FileSystemStorageManager.cpp:
(WebKit::FileSystemStorageManager::releaseLockForFile):
* Source/WebKit/NetworkProcess/storage/StorageAreaBase.cpp:
(WebKit::StorageAreaBase::addListener):
* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::pageEndUsingWebsiteDataStore):