Fix TestWebKitAPI.SiteIsolation.RunOpenPanel https://bugs.webkit.org/show_bug.cgi?id=281276 rdar://137267779
Reviewed by Wenson Hsieh.

It was flaky because of the async nature of WebPageProxy::didChooseFilesForOpenPanel.
It was possible for the evaulateJavaScript message to be received by the web content process
before the WebPage::DidChooseFilesForOpenPanel message.  This makes it so if that happens,
we will just cleanly try again.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, RunOpenPanel)):
(TestWebKitAPI::TEST(SiteIsolation, DISABLED_RunOpenPanel)): Deleted.

Canonical link: https://commits.webkit.org/285026@main