AX: AXIsolatedTree::applyPendingChanges unnecessarily holds a strong-ref to itself https://bugs.webkit.org/show_bug.cgi?id=281361 rdar://137790158
Reviewed by Chris Fleizach.

While it used to be recommended / OK practice to create a `protectedThis` inside function calls, this is no longer
the case. Instead, it's expected that there must be a strong-ref smart pointer somewhere on the stack (this is better
because it reduces ref count churn), which is the case for all callsites of applyPendingChanges. That makes the
`protectedThis` inside `applyPendingChanges` redundant.

* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::applyPendingChanges):

Canonical link: https://commits.webkit.org/285070@main