[CURL] Network sessions are not destroyed in platformFinalize https://bugs.webkit.org/show_bug.cgi?id=281281
Reviewed by Michael Catanzaro.

When we close UIProcess it sends to NetworkProcess message(s)
to destroy all sessions (DestroySession) but right after that
the IPC connection is close which causes that NetworkProcess
can be closed without destroying all sessions.

This change destroys all sessions when the NetworkProcess is
closing.

* Source/WebKit/NetworkProcess/curl/NetworkProcessMainCurl.cpp:

Canonical link: https://commits.webkit.org/285180@main