Fix ASSERT failures in TestWebKitAPI.JavaScriptCore.Incremental*. https://bugs.webkit.org/show_bug.cgi?id=281074 rdar://137526107
Reviewed by NOBODY (OOPS!).

The issue is simply that the IncrementalSweeperMainThread and IncrementalSweeperSecondaryThread
had not initialized JSC before using it.  Added calls to WTF::initializeMainThread() and
JSC::initialize() to take care of this.

Also re-enabled these tests i.e. reverting 284925@main.

Lastly, the test is dependent on GC actually running, so that they can confirm that a destructor
was called.  Since JSGarbageCollect() is not guaranteed to trigger the GC, the tests may sometimes
time out.

Since the goal of the test isn't to test whether the GC gets triggered or not, but that the
IncrementalSweeper gets called by the RunLoop, this patch also changes the tests to call
JSAsynchronousGarbageCollectForDebugging() instead.

JSAsynchronousGarbageCollectForDebugging() is newly introduced, and is the asynchronous version of
JSSsynchronousGarbageCollectForDebugging().  JSSsynchronousGarbageCollectForDebugging() will not
work because it will also synchronously sweep the dead objects at the end of the GC cycle.  As a
result, no sweeping is left for the IncrementalSweeper to do as the test expects.

* Source/JavaScriptCore/API/ExtraSymbolsForTAPI.h:
* Source/JavaScriptCore/API/JSBase.cpp:
(garbageCollectForDebugging):
(JSAsynchronousGarbageCollectForDebugging):
(JSSynchronousGarbageCollectForDebugging):
* Tools/TestWebKitAPI/Tests/JavaScriptCore/JSRunLoopTimer.mm:
(TestWebKitAPI::triggerGC):
(TestWebKitAPI::TEST(JavaScriptCore, IncrementalSweeperMainThread)):
(TestWebKitAPI::TEST(JavaScriptCore, IncrementalSweeperSecondaryThread)):