[Navigation] Improve cross-document traversals
page->goToItem() does not take an arbitary HistoryItem it requires a full HistoryItem tree for the mainFrame so we now make one. Previously the wrong Frame would try to load the item.

Now that the right frame loads the item we also now handle the new entry list for cross-document traverals.