[build.webkit.org] Fix file extension issues when uploading the new bundles after 285255@main https://bugs.webkit.org/show_bug.cgi?id=281671
Unreviewed follow-up after 285255@main.

The new minibrowser bundles are generated with format '.tar.xz' instead of '.zip'
because that generates significant savings on both bandwidth and disk space usage,
but on 285255@main I forgot to update the step that uploads the bundles so it picks
a '.tar.xz' file instead of a '.zip' one. Also the script needs a small modification
to generate the '.sha256sum' file with the right name in this case.

* Tools/CISupport/Shared/transfer-archive-via-sftp:
(SftpClient.upload):
* Tools/CISupport/build-webkit-org/steps.py:
(UploadMiniBrowserBundleViaSftp):
* Tools/CISupport/build-webkit-org/steps_unittest.py:

Canonical link: https://commits.webkit.org/285346@main