Remove ColorInputType::typeMismatchFor https://bugs.webkit.org/show_bug.cgi?id=281620 rdar://138062947
Reviewed by Wenson Hsieh.

110431@main added this for the benefit of filtering <datalist>
suggestions, but these days that is handled by
ColorInputType::suggestedColors which directly parses the result into a
color and therefore takes care of the validation.

And <input type=color> always has a valid value, except when a custom
error is used. So this is not used for that either.

* Source/WebCore/html/ColorInputType.cpp:
(WebCore::ColorInputType::typeMismatchFor const): Deleted.
* Source/WebCore/html/ColorInputType.h:

Canonical link: https://commits.webkit.org/285356@main