Enable setting to block the MobileGestalt service https://bugs.webkit.org/show_bug.cgi?id=281622 rdar://116054248
Reviewed by Sihui Liu.

Enable setting to block the MobileGestalt service in the WebContent process sandbox on iOS.
Local validation has not revealed any issues with this change, so we can move forward with
enabling the setting by default. This patch only enables blocking of the service, not the
reading of the MobileGestalt disk cache, since we are not concerned about read access to
this cache.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb.in:

Canonical link: https://commits.webkit.org/285403@main