Fix TestWebKitAPI.ProcessSwap.ConcurrentHistoryNavigations after 285375@main https://bugs.webkit.org/show_bug.cgi?id=281750
Reviewed by Chris Dumez.

The lifetime of ProvisionalPageProxy is quite sensitive.  In particular, in
WebPageProxy::continueNavigationInNewProcess we set m_provisionalPage to null
before creating a new ProvisionalPageProxy, and the destructor of the old one
needs to be called before creating the new one or we will assert in
WebProcessProxy::addVisitedLinkStoreUser.  285375@main extended the lifetime
of the ProvisionalPageProxy problematically in a few places.  This restores
the old lifetime of the ProvisionalPageProxy without hurting WeakPtr safety.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::receivedNavigationActionPolicyDecision):
(WebKit::WebPageProxy::commitProvisionalPage):

Canonical link: https://commits.webkit.org/285420@main