[WebKitTestRunnerApp] Missing entitlements blocking iOS device install. https://bugs.webkit.org/show_bug.cgi?id=281349 rdar://137778117
Reviewed by Per Arne Vollan.

When WebKit adopted BrowserEngineKit, there were a lot of new app entitlements
that needed to be added to WebKitTestRunnerApp in order to allow it to install
with iOS. These were added to the simulator entitlements, but they were never
added to physical device entitlements. This change adds them for physical
devices.

* Tools/WebKitTestRunner/Configurations/WebKitTestRunnerApp-iOS.entitlements:

Canonical link: https://commits.webkit.org/285452@main